# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Affyne API
  description: |
    REST API for the Affyne application that enables users to register, authenticate, 
    and find music taste matches based on favorite artists.
    
    ## Features
    - User registration with password validation
    - Secure authentication with bcrypt password hashing
    - Artist preference management
    - Music taste matching algorithm with overlap scoring
    - In-memory data storage (suitable for prototyping)
  version: 1.0.0
  contact:
    name: Affyne Team
    url: https://github.com/greenwaltc/kellogg-music-match
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.kellogg-music-match.com
    description: Production server
...
paths:
...
  /push/subscribe:
    post:
      tags:
        - Push
      summary: Store or update a Web Push subscription for the authenticated user
      description: |
        Registers a Web Push subscription for the current user. The server will use this to deliver
        notifications to all of a user's devices. Requires JWT authentication.
      operationId: upsertPushSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushSubscriptionRequest'
            examples:
              typical:
                value:
                  endpoint: "https://fcm.googleapis.com/fcm/send/cR18...:APA91b..."
                  keys:
                    p256dh: "BKNm..."
                    auth: "0o3x..."
      responses:
        '200':
          description: Subscription stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                ok:
                  value: { status: ok }
        '400':
          description: Invalid subscription payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to store subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /push/device:
    delete:
      tags:
        - Push
      summary: Delete a stored native device token
      operationId: deleteDeviceToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platform:
                  type: string
                  enum: [ios, android]
                token:
                  type: string
              required: [platform, token]
      responses:
        '200':
          description: Token deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                deleted:
                  value: { status: deleted }
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to delete token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    StatusResponse:
      type: object
      description: Generic status response
      properties:
        status:
          type: string
          example: ok
      required: [status]
    PushSubscriptionRequest:
      type: object
      description: Standard Web Push subscription payload
      properties:
        endpoint:
          type: string
          format: uri
        keys:
          type: object
          properties:
            p256dh:
              type: string
            auth:
              type: string
          required: [p256dh, auth]
      required: [endpoint, keys]
...
    ErrorResponse:
      type: object
      description: Standard error response format
      properties:
        message:
          type: string
          description: Human-readable error description
          example: "username already exists"
        created_at:
          type: string
          format: date-time
          description: Timestamp when the error occurred
          example: "2025-10-06T12:00:00Z"
      required:
        - message
        - created_at
...
  securitySchemes:
    # JWT implementation now active
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication for secure API access

# Global security - all endpoints except auth require JWT
security:
  - BearerAuth: []
...